Extract Authentication Logic to the Common Project. #81
