+++ Backend/apigateway/src/main/java/com/bikenest/apigateway/UsermgmtClient.java
@@ -12,7 +12,8 @@ import org.springframework.web.bind.annotation.ResponseBody;
 public interface UsermgmtClient {
     @RequestMapping(method = RequestMethod.POST, value = "/usermanagement/validatejwt",
             consumes = {MediaType.APPLICATION_JSON_VALUE},
-            produces = {MediaType.APPLICATION_JSON_VALUE})
+            produces = {MediaType.APPLICATION_JSON_VALUE},
+            headers = {"Authorization: 12345"})
     @ResponseBody
     boolean ValidateJwt(@RequestBody String JWT);
 }
