+++ Backend/apigateway/src/main/java/com/bikenest/apigateway/WebSecurityConfig.java
@@ -1,0 +1,19 @@
+package com.bikenest.apigateway;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.web.server.ServerHttpSecurity;
+import org.springframework.security.web.server.SecurityWebFilterChain;
+
+@Configuration
+public class WebSecurityConfig {
+
+    @Bean
+    public SecurityWebFilterChain springSecurityFilterChain(ServerHttpSecurity http) {
+        return http.authorizeExchange().anyExchange().permitAll()
+            .and()
+            .csrf()
+            .disable()
+            .build();
+    }
+}
