+++ Backend/common/src/main/java/com/bikenest/common/security/AuthToken.java
@@ -1,7 +1,6 @@
-package com.bikenest.servicebikenest.security;
+package com.bikenest.common.security;
 
 import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jws;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 
@@ -25,8 +24,8 @@ public class AuthToken implements Authentication {
     }
 
     @Override
-    public Object getDetails() {
-        return null;
+    public Claims getDetails() {
+        return claims;
     }
 
     @Override
