+++ Backend/common/src/main/java/com/bikenest/common/security/JWTAuthenticationEntrypoint.java
@@ -1,16 +1,15 @@
-package com.bikenest.servicebikenest.security;
+package com.bikenest.common.security;
 
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.io.Serializable;
 
-@Component
-public class JWTAuthenticationEntrypoint implements AuthenticationEntryPoint, Serializable{
+
+public class JWTAuthenticationEntrypoint implements AuthenticationEntryPoint, Serializable {
     private static final long serialVersionUID = -233241356;
     @Override
     public void commence(HttpServletRequest request, HttpServletResponse response,
