+++ Backend/common/src/main/java/com/bikenest/common/security/JWTAuthenticationFilter.java
@@ -1,6 +1,9 @@
-package com.bikenest.servicebikenest.security;
+package com.bikenest.common.security;
 
-import io.jsonwebtoken.*;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Header;
+import io.jsonwebtoken.Jwt;
+import io.jsonwebtoken.Jwts;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.filter.GenericFilterBean;
 
@@ -10,7 +13,8 @@ import javax.servlet.ServletRequest;
 import javax.servlet.ServletResponse;
 import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
-class JWTAuthenticationFilter extends GenericFilterBean {
+
+public class JWTAuthenticationFilter extends GenericFilterBean {
 
     @Override
     public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
@@ -18,7 +22,7 @@ class JWTAuthenticationFilter extends GenericFilterBean {
         String jwt = httpRequest.getHeader("Authorization");
         if(jwt != null){
             String withoutSignature = jwt.substring(0, jwt.lastIndexOf('.')+1);
-            Jwt<Header,Claims> claims = Jwts.parserBuilder().build().parseClaimsJwt(withoutSignature);
+            Jwt<Header, Claims> claims = Jwts.parserBuilder().build().parseClaimsJwt(withoutSignature);
             AuthToken auth = new AuthToken(claims.getBody());
             SecurityContextHolder.getContext().setAuthentication(auth);
         }
